requirements.txt
opencv-python~=4.9.0.80
matplotlib~=3.7.5
simplekml~=1.3.6
sahi~=0.11.19
pillow~=10.4.0
gmplot~=1.4.1
pyproj~=3.5.0